changed int to float to correct bug

+++ cart/views.py
@@ -22,7 +22,7 @@ def add_to_cart(request, id):
 
 def adjust_cart(request, id):
     """ Adjust the quantity of the specified product to the specified ammount"""
-    quantity = int(request.POST.get('quantity'))
+    quantity = float(request.POST.get('quantity'))
     cart = request.session.get('cart', {})
 
     if quantity > 0:
